refactor(RealtimeTimeColumn): Rename component, address PR comments.
