+++ lib/components/narrative/line-itin/connected-itinerary-body.js
@@ -12,7 +12,7 @@ import styled from 'styled-components'
 import { showLegDiagram } from '../../../actions/map'
 import { setViewedTrip } from '../../../actions/ui'
 import TransitLegSubheader from './connected-transit-leg-subheader'
-import TimeColumnWithDelays from './time-column-with-delays'
+import RealtimeTimeColumn from './realtime-time-column'
 import TripDetails from '../connected-trip-details'
 import TripTools from '../trip-tools'
 
@@ -68,7 +68,7 @@ class ConnectedItineraryBody extends Component {
           toRouteAbbreviation={noop}
           TransitLegSubheader={TransitLegSubheader}
           TransitLegSummary={TransitLegSummary}
-          TimeColumnContent={TimeColumnWithDelays}
+          TimeColumnContent={RealtimeTimeColumn}
         />
         <TripDetails itinerary={itinerary} />
         <TripTools itinerary={itinerary} />
